refactor: consolidate schema validation
