+++ tests/test_schema_examples_round_trip.py
@@ -1,4 +1,4 @@
-from parseo.parser import validate_schema_examples
+from parseo import validate_schema_examples
 
 
 def test_schema_examples_round_trip():
